# Enable single-writer multiple-reader access to trace files.