Fixes #224.

By using the latest version of hdf5 we gain access to some extra features, including single-writer multiple-reader (SWMR) mode.
We may wish to specify the exact version we want to use, going forward, to ensure backwards compatibility is more easily assured.

- [x] Test that I can read from a live trace file as it changes.
- [x] Verify the integrity of the traces (compare locally acquired version to version recovered from file)
- [x] I also checked that loading h5py files from previous library versions works well.
- [x] Speed performance, actually seems to have improved
    - Before
        - Storing 600,000 points takes ~ 260 +- 58 ms.
        - Storing 1500 points takes ~5.7 +- 4.7 ms
    - After 
        - Storing 600,000 points takes ~ 230 +- 50 ms.
        - Storing 1500 points takes ~1 +- 4 ms