@maij 2021-11-14
> Great addition! Do we need a minimal h5py version?

Nope! The current minimum in `setup.py` is 2.6.0 which supports SWMR.